rpc-ext: Extract the storage_changes impl into sub mod
